Fixed warning and some cleanups

+++ src/billsdepositspanel.cpp
@@ -695,7 +695,7 @@ void billsDepositsListCtrl::OnEnterBDTransaction(wxCommandEvent& /*event*/)
     mmBDDialog dlg(this, id, false, true);
     if ( dlg.ShowModal() == wxID_OK )
     {
-        if (++m_selected_row < m_bdp->bills_.size())
+        if (++m_selected_row < long(m_bdp->bills_.size()))
             id = m_bdp->bills_[m_selected_row].BDID;
         refreshVisualList(m_bdp->initVirtualListControl(id));
     }
@@ -707,7 +707,7 @@ void billsDepositsListCtrl::OnSkipBDTransaction(wxCommandEvent& /*event*/)
 
     int id = m_bdp->bills_[m_selected_row].BDID;
     Model_Billsdeposits::instance().completeBDInSeries(id);
-    if (++m_selected_row < m_bdp->bills_.size())
+    if (++m_selected_row < long(m_bdp->bills_.size()))
         id = m_bdp->bills_[m_selected_row].BDID;
     refreshVisualList(m_bdp->initVirtualListControl(id));
 }
